+++ packages/oxlint-config-smarthr/CHANGELOG.md
@@ -1,5 +1,61 @@
 # Changelog
 
+## [0.1.8](https://github.com/kufu/tamatebako/compare/oxlint-config-smarthr-v0.1.7...oxlint-config-smarthr-v0.1.8) (2026-06-09)
+
+
+### Features
+
+* oxlint-config-smarthr で eslint-plugin-smarthr を &gt;=6.18.0 に更新 ([#1358](https://github.com/kufu/tamatebako/issues/1358)) ([42d254f](https://github.com/kufu/tamatebako/commit/42d254f6a3353a5e8d4431aa19f9c9c3f17a676a))
+
+## [0.1.7](https://github.com/kufu/tamatebako/compare/oxlint-config-smarthr-v0.1.6...oxlint-config-smarthr-v0.1.7) (2026-06-08)
+
+
+### Features
+
+* eslint-plugin-smarthrを6.17.0に更新 (oxlint-config-smarthr) ([#1346](https://github.com/kufu/tamatebako/issues/1346)) ([2d1ecc7](https://github.com/kufu/tamatebako/commit/2d1ecc79fcd89c9167549e8289cc285484d619be))
+
+## [0.1.6](https://github.com/kufu/tamatebako/compare/oxlint-config-smarthr-v0.1.5...oxlint-config-smarthr-v0.1.6) (2026-06-08)
+
+
+### Bug Fixes
+
+* oxlint-config-smarthr の eslint-plugin-smarthr への依存を dependencies に移動 ([#1338](https://github.com/kufu/tamatebako/issues/1338)) ([2fa5fd7](https://github.com/kufu/tamatebako/commit/2fa5fd7659e4843518aca710d228201d79f0b34b))
+
+## [0.1.5](https://github.com/kufu/tamatebako/compare/oxlint-config-smarthr-v0.1.4...oxlint-config-smarthr-v0.1.5) (2026-05-19)
+
+
+### Features
+
+* eslint-plugin-smarthrを6.16.0に更新 (oxlint-config-smarthr) ([#1326](https://github.com/kufu/tamatebako/issues/1326)) ([e636823](https://github.com/kufu/tamatebako/commit/e6368234c7ad9a6d51b1363ba7d32be9aa49449c))
+
+## [0.1.4](https://github.com/kufu/tamatebako/compare/oxlint-config-smarthr-v0.1.3...oxlint-config-smarthr-v0.1.4) (2026-05-17)
+
+
+### Features
+
+* **oxlint-config-smarthr:** eslint-plugin-smarthrを6.15.0に更新 ([#1317](https://github.com/kufu/tamatebako/issues/1317)) ([713cac4](https://github.com/kufu/tamatebako/commit/713cac4d2e887c302b18e945ea39a62be9a2bb1d))
+
+## [0.1.3](https://github.com/kufu/tamatebako/compare/oxlint-config-smarthr-v0.1.2...oxlint-config-smarthr-v0.1.3) (2026-05-14)
+
+
+### Features
+
+* **oxlint-config-smarthr:** eslint-plugin-smarthrを&gt;=6.14.0に更新 ([#1307](https://github.com/kufu/tamatebako/issues/1307)) ([bcab745](https://github.com/kufu/tamatebako/commit/bcab7456d1456ecca8600d8c0950641669e757fd))
+
+## [0.1.2](https://github.com/kufu/tamatebako/compare/oxlint-config-smarthr-v0.1.1...oxlint-config-smarthr-v0.1.2) (2026-04-19)
+
+
+### Bug Fixes
+
+* update eslint-config-smarthr to 14.0.1 ([#1259](https://github.com/kufu/tamatebako/issues/1259)) ([0f9e2a8](https://github.com/kufu/tamatebako/commit/0f9e2a801c9424f503e3446d9615f9ca7554a474))
+
+## [0.1.1](https://github.com/kufu/tamatebako/compare/oxlint-config-smarthr-v0.1.0...oxlint-config-smarthr-v0.1.1) (2026-04-19)
+
+
+### Bug Fixes
+
+* **oxlint-config-smarthr:** update eslint-plugin-smarthr to &gt;=6.12.1 ([#1257](https://github.com/kufu/tamatebako/issues/1257)) ([fddaca9](https://github.com/kufu/tamatebako/commit/fddaca9535d995388655797c6444e9e5256948af))
+
 ## 0.1.0 (2026-04-14)